umtrx: fix GPS NMEA data, over UART1 (UDP/49171)
ZPU firmware was setup to poll uart2 (== UDP port 47172) on the quad_uart.v
core. And uart2 is connected to RXD[3] and TXD[3] in the top level. However,
the GPS UART lines are connected to RXD[2] and TXD[2] in the UCF -- which
would be UART1 in from the quad_uart.v/firmware's perspective.

This fix makes ZPU firmware use uart1 (UDP/49171) instead of uart2.

(The other alternative was to change this assignment in FPGA firmware:
https://github.com/fairwaves/UHD-Fairwaves/blob/4fa97a15a241/fpga/usrp2/top/N2x0/u2plus_umtrx_v2.v#L631)

Quick way to see GPS NMEA data (GPS antenna should be connected and GPS
LED should blink once a second):

    echo hello | nc -u 192.168.10.2 49171

Huge kudos to @guruofquality for debugging this problem.