fpga: move tx frontend mux between DSP and corrections module
On the RX side, the same mux is between the corrections and the DSP.
However, for TX, this was between the DAC and corrections module,
leading to the host code setting the wrong corrections when the
channel 0 was set to frontend mapping "B:0".

This fixes the issue with the tx iq calibration utility.
A new fpga image is checked in and compat minor bumped.