host: Check if DC offset value reads as it's been written to LMS.
We had an issue with GPS 1pps signals disrupting SPI operations,
which was mostly visible during calibratoin, because it takes
many SPI operations to finish. And because the key operation
during calibration is DC offset altering, it's easy to spot when
SPI is working incorrectrly by checking it. It's a useless check
in a normal situation, so I don't want to see this in production
code.